renaming auth interface file to auth-interface
